add first test case using London School
