+++ Java/src/main/java/TicketOffice.java
@@ -1,11 +1,17 @@
 public class TicketOffice {
     
+    private TrainDataService trainDataService;
+
     public TicketOffice(String trainDataService, String bookingReferenceService) {
 		//TODO: implement this code!
     }
 
+    public TicketOffice(TrainDataService trainDataService, Object bookingReferenceService) {
+        this.trainDataService = trainDataService;
+    }
+
     public Reservation makeReservation(ReservationRequest request) {
-		//TODO: implement this code!
+        trainDataService.getTrainInformation("express_2000");
 		return null;
     }
