+++ Java/src/test/java/TrainDataService.java
@@ -1,0 +1,6 @@
+
+public interface TrainDataService {
+
+    void getTrainInformation(String trainId);
+
+}
